Commands as format strings, see #137
This commit adds dox, tests and of course the changes as discussed in
issue #137 to use format strings as a standard for the run() function.

It also updates ctx.run() to work by defaulting the expansion vars in
the run() function, with explicit overrides and additional values
available in the expand kwarg.